api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.3
httpx==0.26.0
pypdf==4.0.1
python-docx==1.1.0
pandas==2.2.0
qdrant-client==1.7.3
tiktoken==0.5.2
passlib==1.7.4
python-jose==3.3.0
cryptography==42.0.0
langwatch==0.0.4
